First pass at e-mailing bookmarks.
Removed our custom controller and views and let the default
Blacklight views/controllers handle most of the code for
Bookmarks.

The code to e-mail bookmarks was revamped to include Location
and Call Number. There is some logic to defice what to include
on the e-mail as we don't have a good way to determine what
Location correspond to each Call Number. The current version
works OK for items with only one Location and one Call Number.

Fixed Blacklight bug that only e-mails users the firt 10
bookmarks.

Created a Location model and table to handle Library Locations.
Added rake task to initialize the database with a list of
locations.