Fixes issue in Search Result that caused the "Item" link to be rendered with an invalid URL.
Fixed JavaScript code in catalog_results_availability.js
to make the correct validations on when to show the link
and set the appropriate URL for it.

I also added a brand new Availability controller to allow
us to bypass the AJAX same-domain security while testing
the system in the development server. This allow us to
execute the code that parses the Availability Service
and make sure the pages will work when deployed to production.