Started switching the virtual shelf to load data via AJAX calls.
Still pending to allow the user to fetch more books with the up/down arrows
but it seems that this is not supported out of the box, instead the Stackview
expects an infinite scroll feature via a mouse plugin. Need to look more into
this.

There still work to do in the API controller to clean up the nearby method.
Right now I have two versions but I only need one.